chore(tests): move ExtractRemainingFromTableRuntime to plugin_collection
