+++ falco_plugin_tests/src/plugin_collection/extract/mod.rs
@@ -1,1 +1,2 @@
 pub mod remaining_from_table;
+pub mod remaining_from_table_runtime;
